% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dispersion-calculator-package.r
\docType{package}
\name{DispersionCalculator}
\alias{DispersionCalculator}
\alias{DispersionCalculator-package}
\title{DispersionCalculator}
\description{
DispersionCalculator
}

